[ADD] ge03_team01: compute the product name for motorcycles
Kawiil needed to automatically name their motorcycle products based on the make, model
and year of the product. This module implements a computed field on the 'product.template'
model that automatically assigns the name to products with 'detailed_type' equal to
'motorcycle', based on its make, model and year fields.