[IMP] ge08_team01: 'Motorcycle Registries' portal views
Kawiil wanted their customers to view their registries information on the customer portal.
They as well wanted their customers to be able to edit and to opt in to make their records
public and to see others records.

This module adds 'Motorcycle Registries' to the 'My Account' page. This controller redirects
to the registries list view. All of the user's registries as well as public ones (the ones
with the field 'is_public' set to True) are displayed here. Users may filter registries by
state, country or owner. Each registry in the list has a link button in the 'registry_number',
which redirects the user to the registry's form view.

A 'Back to edit mode' was added in the form view, which redirects the user to the registry's
backend view, to let them edit their own registries.

Handle search by make and model

[IMP] Removed unnecesary comment