[ADD] ge07_team01: create registry when delivery is validated
Kawiil wanted to automate the process of creating a Motorcycle
Registry when a product is delivered. This module implements the
automatic creation of a Motorcycle Registry when an order deliver
(that has at least one product with a detailed type of 'motorcycle')
is validated.

A Lot ID field is added to the 'motorcycle.registry' model, which is
displayed in its list view. A Registry ID was also added to the 'stock.
lot' model, with the same behavior. Users are also able to track the
sale order of a Motorcycle Registry via the smart button in the form
view of this model.

task-3427405