[ADD] ge06_team01: VIN as the serial number for motorcycles
Kawiil didn't want to confuse its employees with both a serial number
(in the manufacturing order) and a VIN (in the registry). Therefore,
they wanted these two fields to work as one for their motorcycles.

This module assigns a serial number to the products with 'motorcycle'
as its 'detailed_type' following the next rules:
- Make: 2 capital letters
- Model: 2 capital letters
- Year: 2 numbers
- Sequence: 6 numbers