[IMP] Added the 'Motorcycle Registries' portal views
This module adds a 'Motorcycle Registries' button to the 'My Account' page. This button
redirects the user to the registries' list view. All of the user's registries will be
displayed there, as where as other users registries that have the 'is_public' field set
to True. Users may filter registries by state, country or owner. Each registry in the
list has a link button in the 'registry_number', which redirects the user to the registry's
form view.

This view as well has a link button in each registry number, which redirects the user
to the registry's form view. A 'Back to edit mode' was added in this view, which redirects
the user to the registry's backend view, to let them edit their own registries.