[ADD] ge11_team01 module that allows customers to create maintenance tickets online without any external intervention. 
I start creating a new model called 'repair' that inherit from "repair.order". There I defined some
functions and then I create new controllers for managing my portal views where the customers will
create and view their repair orders.
Once we finish defining all the functions that will help us prepare and render data, we start working
with views. There we have to manage a lot of fields that we want to display for the user as well as a
few bootstrap fields to make beautiful views.
To end we create a rause UserError just to ensure that the user will enter a valid/existing VIN.