[ADD] ge10_team01: added widget for total mileage
The customer, Kawiil, wanted but wasn't able to showcase the total mileage of all of their sold motorcycles.
This module adds a widget available through the webside editor. The widget displays the summation of all the
'current_mileage' fields of the 'motorcycle.registry' model. Users may customize this widget with the next
attributes: layout; box or text, background color, border color, text color.